# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Discord Bot Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: Get guilds the user manages
      responses:
        "200":
          description: List of guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"
        "401":
          description: Unauthorized
  /guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get a specific guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guild"
        "404":
          description: Not found
  /guilds/{guildId}/channels:
    get:
      operationId: getGuildChannels
      tags: [guilds]
      summary: Get guild channels
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
  /guilds/{guildId}/roles:
    get:
      operationId: getGuildRoles
      tags: [guilds]
      summary: Get guild roles
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
...
  /guilds/{guildId}/logs:
    get:
      operationId: getGuildLogs
      tags: [logs]
      summary: Get recent activity logs for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
  /guilds/{guildId}/stats:
    get:
      operationId: getGuildStats
      tags: [stats]
      summary: Get dashboard stats for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildStats"
components:
  schemas:
...
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        botPresent:
          type: boolean
        memberCount:
          type: ["integer", "null"]
      required:
        - id
        - name
        - botPresent
    Channel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: integer
      required:
        - id
        - name
        - type
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: integer
        position:
          type: integer
      required:
        - id
        - name
        - color
        - position
...
    ActivityLog:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        type:
          type: string
        userId:
          type: ["string", "null"]
        targetId:
          type: ["string", "null"]
        details:
          type: string
        createdAt:
          type: string
      required:
        - id
        - guildId
        - type
        - details
        - createdAt
    GuildStats:
      type: object
      properties:
        guildId:
          type: string
        totalMembers:
          type: integer
        totalEmbeds:
          type: integer
        totalModActions:
          type: integer
        totalJoinsToday:
          type: integer
        totalLeavesToday:
          type: integer
        recentLogs:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
      required:
        - guildId
        - totalMembers
        - totalEmbeds
        - totalModActions
        - totalJoinsToday
        - totalLeavesToday
        - recentLogs